Move `Message` to top level
