+++ src/pipeline/mod.rs
@@ -29,14 +29,15 @@ mod server;
 mod pipeline;
 
 pub use self::server::Server;
-
 pub use self::client::{connect, Client, Connecting};
+
+use {Message};
 use tokio_core::io::FramedIo;
 use tokio_service::Service;
 use futures::{Async, Future, IntoFuture, Poll};
 use futures::stream::{Stream, Sender};
 use take::Take;
-use std::{cmp, fmt, io, ops};
+use std::{fmt, io};
 
 /// A pipelined protocol frame
 pub enum Frame<T, B, E> {
@@ -53,14 +54,6 @@ pub enum Frame<T, B, E> {
     Error(E),
     /// Final frame sent in each transport direction
     Done,
-}
-
-/// Message sent and received from a pipeline service
-pub enum Message<T, B> {
-    /// Has no associated streaming body
-    WithoutBody(T),
-    /// Has associated streaming body
-    WithBody(T, B),
 }
 
 /// A specialization of `Service` supporting the requirements of server
@@ -221,75 +214,6 @@ impl<T, B, E> fmt::Debug for Frame<T, B, E>
             Frame::Body(ref v) => write!(fmt, "Frame::Body({:?})", v),
             Frame::Error(ref v) => write!(fmt, "Frame::Error({:?})", v),
             Frame::Done => write!(fmt, "Frame::Done"),
-        }
-    }
-}
-
-/*
- *
- * ===== impl Message =====
- *
- */
-
-impl<T, B> Message<T, B> {
-    /// If the `Message` value has an associated body stream, return it. The
-    /// original `Message` value will then become a `WithoutBody` variant.
-    pub fn take_body(&mut self) -> Option<B> {
-        use std::ptr;
-
-        // unfortunate that this is unsafe, but I think it is preferable to add
-        // a little bit of unsafe code instead of adding a useless variant to
-        // Message.
-        unsafe {
-            match ptr::read(self as *const Message<T, B>) {
-                m @ Message::WithoutBody(..) => {
-                    ptr::write(self as *mut Message<T, B>, m);
-                    None
-                }
-                Message::WithBody(m, b) => {
-                    ptr::write(self as *mut Message<T, B>, Message::WithoutBody(m));
-                    Some(b)
-                }
-            }
-        }
-    }
-}
-
-impl<T, B> cmp::PartialEq<T> for Message<T, B>
-    where T: cmp::PartialEq
-{
-    fn eq(&self, other: &T) -> bool {
-        (**self).eq(other)
-    }
-}
-
-impl<T, B> ops::Deref for Message<T, B> {
-    type Target = T;
-
-    fn deref(&self) -> &T {
-        match *self {
-            Message::WithoutBody(ref v) => v,
-            Message::WithBody(ref v, _) => v,
-        }
-    }
-}
-
-impl<T, B> ops::DerefMut for Message<T, B> {
-    fn deref_mut(&mut self) -> &mut T {
-        match *self {
-            Message::WithoutBody(ref mut v) => v,
-            Message::WithBody(ref mut v, _) => v,
-        }
-    }
-}
-
-impl<T, B> fmt::Debug for Message<T, B>
-    where T: fmt::Debug
-{
-    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Message::WithoutBody(ref v) => write!(fmt, "Message::WithoutBody({:?})", v),
-            Message::WithBody(ref v, _) => write!(fmt, "Message::WithBody({:?}, ...)", v),
         }
     }
 }
